move some tests to server.test.js
